requirements.txt
specktre==0.2.0
Pillow==5.0.0
git+https://github.com/Rapptz/discord.py.git@async